Change style to match ÖAW color
Change day highlighting back to a color underline
with two different degrees of line thickness
depending on the number of underlying events:
Thin line if there is one event, thick line if there is more
than one event.

Change colors of the highlight and of the hover effect
in popover elements to match the hue of the ÖAW logo.