Fix overflow issues
Month containers have a fixed pixel height set in the
year-calendar.css file, which causes the individual months
to appear smushed together in some years. The quick fix
is to just increase the fixed height from 200px to 230px.

Set overflow to visible for the calendar, so there should
never be a scroll bar within the page.

Reduce box shadow size for dates with several events
per day from 8px to 6px.