Add detailed view for selected calendar events
Use a jqery datatable to display all events in the calendar
within a selected date range.
A date range can be selected via drag and drop
in the calendar view.

Selecting a date range without any events inside
hides the data table.

Still missing: Information about "Abt." and "Band."
of the individual events, as well as structured information
about the events. For now the "event.name" is displayed
in the column "Titel / Sitzung / Tagesordnung"
as raw text without any structure or HTML markup.